create the launcher class
